requirements.txt
Django==4.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
python-dotenv==0.20.0
PyJWT==2.4.0
bcrypt==3.2.2

